% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plungr.R
\name{create_plungr}
\alias{create_plungr}
\title{Create a package for a Plumber API using plungr}
\usage{
create_plungr(
  path,
  check_name = TRUE,
  open = TRUE,
  package_name = basename(path)
)
}
\arguments{
\item{path}{Name of the folder to create the package in. This will also be
used as the package name.}

\item{check_name}{When using this function in the console, you can prevent
the package name from being checked.}

\item{open}{Boolean open the created project}

\item{package_name}{Package name to use by default it's `basename(path)` but if path == '.' and `package_name`
not explicitly given, then `basename(getwd())` will be used.}
}
\description{
Create a package for a Plumber API using plungr
}
